\encoding{utf8}
\name{adsale}
\docType{data}
\alias{adsale}

\title{
  adsale
}

\description{
 일반적으로 우리는 광고비의 투자가 많을수록 판매액이 증가할 것이라
 생각할 수 있다. 자료 \code{adsale}는 이를 구체적으로 조사하기 위하여
 같은 제품을 만다는 회사들에 대한 판매액과 광고비를 조사한 것이다.
}

\usage{adsale}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{sale}{판매액(백만원)}
    \item{ad}{광고비(백만원)}
    \item{media}{광고매체}
    }
}

\references{
}

\source{
}

\keyword{datasets}