\encoding{utf8}
\name{blood2001m}
\docType{data}
\alias{blood2001m}

\title{
  blood2001m
}

\description{
 자료 \code{blood2001m}는 보건복지부 산하 질병본부에서 2001년에 조사한
 우리나라 남자들의 나이별 평균 혈압이다.
}

\usage{blood2001m}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{bp}{혈압}
    \item{age}{나이}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
