\encoding{utf8}
\name{bloodpressure}
\docType{data}
\alias{bloodpressure}

\title{
  bloodpressure
}

\description{
  자료 \code{bloodpressure}는 20세에서 60세 사이의 건강한 여성 54명에
  대해 이완기 혈압과 나이를 조사한 것이다.


}

\usage{bloodpressure}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{bp}{혈압}
    \item{age}{나이}
    }
}

\examples{
}

\references{
  Neter et al. 1996
}

\source{
}

\keyword{datasets}
