\encoding{utf8}
\name{bod}
\docType{data}
\alias{bod}

\title{
  bod
}

\description{
  자료 \code{bod}는 개울물의 생화학적 산소요구량(BOD)을 측정한 것이다.

}

\usage{bod}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{BOD}{생화학적 산소요구량}
    \item{day}{시간(날짜)}
    }
}

\examples{
}

\references{

}

\source{
}

\keyword{datasets}
