\encoding{utf8}
\name{bodyfat}
\docType{data}
\alias{bodyfat}

\title{
  bodyfat
}

\description{
\code{bodyfat}
}

\usage{bodyfat}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{fat}{체내 지방량}
    \item{x1}{삼두근 두께}
    \item{x2}{허벅지 둘레}
    \item{x3}{팔의 중간부위 둘레}
    }
}

\references{
  Neter et al. 1996
}

\source{

}

\keyword{datasets}
