\encoding{utf8}
\name{bookprice}
\docType{data}
\alias{bookprice}

\title{bookprice}

\description{
  책의 가격을 결정하는 과정에서 책의 쪽수는 중요한 요인으로 작용한다.
  당연히 쪽수가 많을 수록 책의 가격은 높아질 것이라 생각할 수 있다. 자료
  \code{bookprice}는 이를 구체적으로 분석하기 위하여 1997년 6월에 통계학
  관련 전공서적의 쪽수와 가격을 조사한 것이다.
 }

\usage{bookprice}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{price}{가격(천원)}
    \item{pages}{쪽수}
    }
}

\references{
}

\source{
}

\keyword{datasets}