\encoding{utf8}
\name{bug}
\docType{data}
\alias{bug}

\title{
  bug
}

\description{
자료 \code{bug}는 살충제의 독성실험에서 살충제에 노출된 벌레들의
생존개체수를 시간대별로 관측한 것이다. 
}

\usage{bug}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{y}{생존벌레의 수}
    \item{time}{시간(분)}
    }
}

\references{
}

\source{
}

\keyword{datasets}
