\encoding{utf8}
\name{carsale}
\docType{data}
\alias{carsale}

\title{carsale}

\description{ 자료 \code{carsale}은 26개 자동차 영업소에서 1년 동안
  자동차 판매실적을 나타낸 것으로 총판매액, 광고비, 영업장 면적, 차량
  종류, 영업사원수의 값이 주어져 있다.}

\usage{carsale}

\format{이 자료는 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{sale}{총판매액(억원)}
    \item{ad}{광고비(백만원)}
    \item{area}{영업장 면적(평방미터)}
    \item{car}{차량 종류}
    \item{emp}{영업사원수}
    }
}


\examples{
}

\references{
}
\source{
}
