\encoding{utf8}
\name{catalyst}
\docType{data}
\alias{catalyst}

\title{
  catalyst
}

\description{
  어느 화학공장에서 수율을 최대화하는 촉매의 양을 찾기 위하여 현재
  사용하고 있는 촉매의 농도 근처에서 실험을 하여 \code{catalyst}의 결과를 얻었다.
}

\usage{catalyst}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{yield}{수율}
    \item{conc}{촉매의 농도}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
