\encoding{utf8}
\name{chemical}
\docType{data}
\alias{chemical}

\title{
  chemical
}

\description{
  자료 \code{chemical}은 시간에 따른 A와 B의 농도 자료이다.
  A라는 화학물질이 반응을 일으켜 B라는 물질로 변해간다고 하고,
  반응속도는 현재 남아 있는 A의 농도에 비례한다고 가정하자. 이를
  일차반응 속도식(first-order kinetics)이라 한다. 이 화학반응 시스템은
  다음과 같은 미분방적식으로 표현될 수 있다.


}

\usage{chemical}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{time}{시간}
    \item{a}{A의 농도}
    \item{b}{B의 농도}
    
    }
}

\examples{
}

\references{

}

\source{
}

\keyword{datasets}
