% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic.R
\name{dbvnorm}
\alias{dbvnorm}
\title{The Bivariate Normal Distribution}
\usage{
dbvnorm(x, mean = c(0, 0), sd = c(1, 1), rho = 0)
}
\arguments{
\item{x}{a 2-dimensional vector or a 2 by n matrix}

\item{mean}{a mean vector}

\item{sd}{a standard deviation vector}

\item{rho}{correlation coefficient}
}
\description{
The Bivariate Normal Distribution
}
\examples{
x <- seq(-3,3,length.out=50)
y <- seq(-3,3,length.out=50)
z <- outer(x, y, function(x, y) dbvnorm(cbind(x, y), rho=0.5))
persp(x,y,z,theta=5,phi=20,col="lightblue", expand=0.5, zlab="density")
contour(x,y,z, xlab="x", ylab="y")
}

