\encoding{utf8}
\name{english1}
\docType{data}
\alias{english1}

\title{
  english1
}

\description{ 어느 4년제 대학교에서 학년 별 영어실력을 비교하기 위하여
  각 학년에서 6명씩 무작위로 추출하여 같은 영어시험을 보게 한 결과 자료
  \code{english1}을 얻었다. 3학년에서 1명, 4학년에서 2명의 결시자가
  있었다.
}

\usage{english1}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{score}{점수}
    \item{grade}{학년}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
