\encoding{utf8}
\name{english2}
\docType{data}
\alias{english2}

\title{
  english2
}

\description{
최근에 제안된 영어교육방법 A, B, C를 비교하기 위해 6개월 동안 각
영어교육방법으로 교육을 받은 후 영어시험성적을 측정하였다. 이런 경우
영어성적에 영향을 주는 요인들은 교육방법 외에도 여러 가지가 있는데
중학교 2학년 남학생으로 교육대상자를 정하면 나이와 성별은 고정되며,
우리가 조정하기 힘든 요인들 중에서 각 교육대상자의 IQ는 랜덤화로
평준화시키고, 교육 전 영어실력은 사전에 비슷한 영어시험으로 측정하여
모형에 포함시킬 수 있다. 자료 \code{english2}에는 각 그룹별로 9명씩 전체
27명의 교육 후 영어성적, 교육전 영어성적, 교육방법종류가 있다.
}

\usage{english2}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{postscore}{교육 후 영어성적}
    \item{prescore}{교육 전 영어성적}
    \item{method}{교육 방법}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
