\encoding{utf8}
\name{enzyme}
\docType{data}
\alias{enzyme}

\title{
  효소의 반응 속도
}

\description{
  자료 \code{enzyme}은 배양액의 농도에 따른 효소의 반응속도를 측정한
  자료이다. 
}

\usage{enzyme}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{rate}{반응 속도}
    \item{conc}{농도}
    }
}

\references{
}

\source{
}

\keyword{datasets}
