\encoding{utf8}
\name{examscore}
\docType{data}
\alias{examscore}

\title{examscore}

\description{ 
자료 \code{examscore}는 어느 대학교 회귀분석 과목의 중간고사와 기말고사 점수이다.  }

\usage{examscore}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{mid}{중간고사 점수}
    \item{term}{기말고사 점수}
    }
}


\references{
}

\source{
}

\keyword{datasets}

