\encoding{utf8}
\name{expdata1}
\docType{data}
\alias{expdata1}

\title{expdata1}

\description{ 자료 \code{expdata1}에는 변수 y, x1, x2가 순서대로
  저장되어 있다. }

\usage{expdata1}

\format{이 자료는 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{y}{반응변수}
    \item{x1}{설명변수}
    \item{x2}{설명변수}
    }
}


\examples{
}

\references{
}
\source{
}