\encoding{utf8}
\name{factorial}
\docType{data}
\alias{factorial}

\title{
  factorial
}

\description{
\code{factorial}
}

\usage{factorial}

\format{
    \describe{
    \item{y}{반응 변수}
    \item{x1}{설명 변수}
    \item{x2}{설명 변수}
    \item{x3}{설명 변수}
    }
}

\references{
}

\source{

}

\keyword{datasets}
