% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slr.R
\name{fitplot}
\alias{fitplot}
\title{Fitted Objectel Plot for Simple Linear Regression}
\usage{
fitplot(object, level = 0.95, legend = c("none", "topleft"))
}
\arguments{
\item{object}{lm object. simple linear regression.}

\item{level}{tolerance/confidence level}

\item{legend}{legend location. See \code{\link[graphics]{legend}}}
}
\description{
Fitted Objectel Plot for Simple Linear Regression
}

