\encoding{utf8}
\name{gdpgrowth}
\docType{data}
\alias{gdpgrowth}

\title{
  gdpgrowth
}

\description{
  자료 \code{gdpgrowth}는 61개 국가에 대해 1960년부터 1985년까지의
  경제성장변수들을 조사한 것이다.
}

\usage{gdpgrowth}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{country}{국가명}
    \item{gdp}{국내총생산 증가(growth in grosss deomestic product per worker)}
    \item{lfg}{노동력 증가(labor force growth)}
    \item{eqp}{설비투자(equipment investment)}
    \item{neq}{비설비투자(nonequipment investment)}
    \item{gap}{relative GDP gap}
    }
}

\examples{
}

\references{
  De Long & Summers, 1991
}

\source{
}

\keyword{datasets}
