\encoding{utf8}
\name{growth2}
\docType{data}
\alias{growth2}

\title{
  growth2
}

\description{
  자료 \code{growth2}는 시간에 따른 식물의 성장치를 기록한 것이다.

}

\usage{growth2}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{growth}{성장치}
    \item{time}{시간}
    }
}

\examples{
}

\references{

}

\source{
}

\keyword{datasets}
