\encoding{utf8}
\name{hald}
\docType{data}
\alias{hald}

\title{
  hald
}

\description{ 자료 \code{hald}는 1932년 미국 화학공학 학회지에 발표된
것으로 시멘트가 굳을 때 발생하는 열량에 시멘트의 네 가지 성분들이 미치는
영향을 분석한 것이다. 
}

\usage{hald}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{y}{시멘트 1g당 발생하는 열량(cal)}
    \item{x1}{성분1의 무게의 구성비(\%)}
    \item{x2}{성분2의 무게의 구성비(\%)}
    \item{x3}{성분3의 무게의 구성비(\%)}
    \item{x4}{성분4의 무게의 구성비(\%)}
    }
}

\references{
  Hald, 1952
  
  Woods et al, 1932
  
  Drapper & Smith, 1998
}

\source{
}

\keyword{datasets}
