% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic.R
\name{histf}
\alias{histf}
\title{Fancy Histogram}
\usage{
histf(x, breaks = sqrt(length(x)), length.out = 100, rug = FALSE,
  boxplot = FALSE, fill = "lightblue", main = paste("Histogram of",
  xname), densityFunction = function(y) dnorm(y, mean(x), sd(x)),
  densityName = "normal", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{breaks}{see \link[graphics]{hist}}
}
\description{
Fancy Histogram
}
\examples{
with(subset(hweight, gender=="M"), histf(height))
with(subset(hweight, gender=="M"), {
   hist(height, probability=TRUE, breaks=sqrt(length(height)), col="lightblue")
   lines(density(height))
   xs <- seq(min(height), max(height), length.out=100)
   lines(xs, dnorm(xs))
})
}

