\encoding{utf8}
\name{houseprice}
\docType{data}
\alias{houseprice}

\title{houseprice}

\description{ 자료 \code{houseprice}은 주택 판매가격과 이에 영향을 줄
  것으로 판단되는 4가지 설명변수인 세금, 대지평수, 건물평수, 주택연령을
  27가구에 대해 조사한 것이다.}

\usage{houseprice}

\format{이 자료는 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{houseprice}{주택 판매가격(천만원)}
    \item{tax}{세금(만원)}
    \item{ground}{대지평수(평)}
    \item{floor}{건물평수(평)}
    \item{year}{주택연령(년)}
    }
}


\examples{
}

\references{
}
\source{
}
