\encoding{utf8}
\name{KBO1}
\docType{data}
\alias{KBO1}

\title{KBO1}

\description{
자료  \code{KBO1}는 2011년 1군 경기에서 1경기 이상 뛴 타자 162명의
선수들을 대상으로 조사한 것이다. 
}

\usage{KBO1}

\format{이 자료는 다음과 같은 변수로 이루어져 있다.
  \describe{
    \item{Y}{연봉(천만원)}
    \item{X1}{안타수}
    \item{X2}{홈런수}
    \item{X3}{타점수}
    \item{X4}{득점수}
    \item{X5}{볼넷수}
    \item{X6}{통산안타수}
    \item{X7}{통산경력}
    \item{team}{구단명}
    \item{player}{선수명}
    }
}

\references{
}
\source{
  http://www.koreabaseball.com
}
