% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diag.R
\name{lmbeta}
\alias{lmbeta}
\alias{lmbeta.lm}
\alias{print.lmbeta}
\alias{summary.lmbeta}
\title{Add Standardized Coefficients}
\usage{
lmbeta(mod, ...)

\method{lmbeta}{lm}(mod, ...)

\method{print}{lmbeta}(obj, ...)

\method{summary}{lmbeta}(obj, ...)
}
\arguments{
\item{mod}{}

\item{...}{}
}
\description{
Add Standardized Coefficients
}

