\encoding{utf8}
\name{mammal}
\docType{data}
\alias{mammal}

\title{mammal}

\description{
자료 \code{mammal}은 포유류들의 뇌의 무게와 몸무게를 조사한 것이다. 전체
자료수는 62개이며 2개의 변수를 가지고 있다.
}

\usage{mammal}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{brain}{뇌의 무게(g)}
    \item{body}{몸무게(kg)}
    }
}

\references{
  Weisberg. 1985.
}

\source{
}

\keyword{datasets}