% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diag.R
\name{meanvar}
\alias{meanvar}
\alias{plot.meanvar}
\alias{print.meanvar}
\title{Mean and Variance of Data Subsets}
\usage{
meanvar(formula, data, ..., na.action = na.omit)

\method{print}{meanvar}(obj, ...)

\method{plot}{meanvar}(obj, formula, ...)
}
\arguments{
\item{formula}{a formula, such as \code{y ~ x}, where the \code{y} variable are numeric data to be split into groups according to the grouping \code{x} variables.}

\item{data}{a data frame}

\item{obj}{an \code{meanvar} object}

\item{formula}{mean-vairance relationship, for example, \code{sd ~ mean}, \code{sd ~ mean + I(mean^2)}, \code{var ~ mean}, etc.}
}
\description{
Splits the data into subsets according to the X levels, compute mean and
variance of the response for each.
}
\examples{
meanvar(Y ~ X, restaurant)
plot(meanvar(Y ~ X, restaurant), sd ~ mean + I(mean^2))
}

