\encoding{utf8}
\name{MLB1}
\docType{data}
\alias{MLB1}

\title{
  MLB1
}

\description{
  자료 \code{MLB1}은 미국 프로야구의 아메리칸리그 소속 타자 124명에 대한
  1985년과 1986년 타율을 나타낸 것이다.
}

\usage{MLB1}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{hit85}{1985년 타율}
    \item{hit86}{1986년 타율}
    }
}

\examples{

}

\references{
}

\source{
}

\keyword{datasets}
