\encoding{utf8}
\name{normaldensity}
\docType{data}
\alias{normaldensity}

\title{
  normaldensity
}

\description{
 자료 \code{normaldensity}는 표준정규분포의 확률밀도함수에 적당한 크기의
 오차를 더하여 생성한 자료이다.

}

\usage{normaldensity}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{y}{반응 변수}
    \item{x}{설명 변수}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
