% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr.fit}
\alias{pcr.fit}
\title{Fitting Principal Component Regression Model}
\usage{
pcr.fit(x, y, ncomp, cor = TRUE, ...)
}
\arguments{
\item{x}{explanatory matrix}

\item{y}{response vector}

\item{ncomp}{number of principal components}

\item{cor}{a logical value indicating whether the calculation should use
the correlation matrix or the covariance matrix. See \code{\link[stats]{princomp}}}

\item{...}{}
}
\description{
Fitting Principal Component Regression Model
}

