% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr}
\alias{pcr}
\alias{print.pcr}
\alias{summary.pcr}
\title{Principal Component Regression}
\usage{
pcr(formula, data, ncomp, cor = TRUE, ...)

\method{summary}{pcr}(object, ...)

\method{print}{pcr}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\description{
Principal Component Regression
}

