% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modsel.R
\name{press.regsubsets}
\alias{press.regsubsets}
\alias{print.press.regsubsets}
\title{Compute Predicted Residual Sum of Squares}
\usage{
\method{press}{regsubsets}(object, ...)

\method{print}{press.regsubsets}(object, ...)
}
\arguments{
\item{object}{a \code{\link[leaps]{regsubsets}} object}

\item{...}{}

\item{...}{}
}
\description{
Compute Predicted Residual Sum of Squares
}

