\encoding{utf8}
\name{quadratic}
\docType{data}
\alias{quadratic}

\title{quadratic}

\description{
자료 \code{quadratic}은 두 변수 x, y에 대한 자료이다. 자료는 두 변수간의
정확한 이차함수관계에 정규분포에서 생성한 난수를 더하여 만들어졌다. 
 }

\usage{quadratic}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{x}{0.1에서 2.0까지 0.1 간격으로 생성.}
    \item{y}{정확한 이차함수관계에 정규분포에서 생성한 난수를 더하여 생성.}
    }
}

\references{
}

\source{
}

\keyword{datasets}