\encoding{utf8}
\name{restaurant}
\docType{data}
\alias{restaurant}

\title{restaurant}

\description{
자료 \code{restaurant}는 비슷한 업종의 음식점 30개의 매출액과 광고비를
조사한 것이다.
}

\usage{restaurant}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{Y}{매출액}
    \item{X}{광고비}
    }
}

\references{
  Weisberg. 1985.
}

\source{ ... }

\keyword{datasets}