% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modsel.R
\name{rpredict}
\alias{rpredict}
\alias{rpredict.lm}
\title{Predicted Residuals}
\usage{
rpredict(model, ...)

\method{rpredict}{lm}(model, infl = lm.influence(model, do.coef = FALSE))
}
\arguments{
\item{model}{an R object, typically by \code{lm}}

\item{...}{further arguments pass to or from other methods}

\item{infl}{influence structure}
}
\description{
Predicted Residuals
}

