\encoding{utf8}
\name{rsadata1}
\docType{data}
\alias{rsadata1}

\title{
  rsadata1
}

\description{
자료 \code{rsadata1}는 설명변수를 x1, x2를 이용한 실험에서 반응변수의
값을 측정한 것이다.
}

\usage{rsadata1}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{y}{반응 변수}
    \item{x1}{설명 변수}
    \item{x2}{설명 변수}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
