% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diag.R
\name{stdcoef}
\alias{stdcoef}
\alias{stdcoef.lm}
\title{Extract Standardized Coefficients}
\usage{
stdcoef(mod, ...)

\method{stdcoef}{lm}(mod, ...)
}
\arguments{
\item{mod}{}

\item{...}{}
}
\description{
Extract Standardized Coefficients
}

