% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modsel.R
\name{summaryf.regsubsets}
\alias{print.summaryf.regsubsets}
\alias{summaryf.regsubsets}
\title{Alternative summary for regsubsets object}
\usage{
\method{summaryf}{regsubsets}(object, ...)

\method{print}{summaryf.regsubsets}(object, ...)
}
\arguments{
\item{object}{a \code{\link[leaps]{regsubsets}} object}

\item{...}{}
}
\description{
Alternative summary for regsubsets object
}

