\encoding{utf8}
\name{suneung}
\docType{data}
\alias{suneung}

\title{suneung}

\description{ 자료 \code{suneung}은 고등학교 3학년 학생 25명의 수학능력
  시험점수와 4개 교과목의 학교성적을 조사한 것이다. }

\usage{suneung}

\format{이 자료는 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{suneung}{수능점수}
    \item{kor}{국어}
    \item{eng}{영어}
    \item{math}{수학}
    \item{sci}{과학}
    }
}


\examples{
}

\references{
}
\source{
}
