\encoding{utf8}
\name{tree}
\docType{data}
\alias{tree}

\title{tree}

\description{ 산림지역에서 나무를 벌목할 때 해당 지역의 목재량을 조사할
  필요가 있다. 그러나 나무의 부피를 직접 측정하는 것은 어렵기 때문에
  상대적으로 측정이 쉬운 나무의 지름과 높이를 이용하여 부피를 추정하는
  방법을 생각할 수 있다. 자료 \code{tree}는 어느 지역에서 15그루의
  나무를 표본으로 추출하여 벌목한 후, 부피, 지름, 높이를 측정한 값들이다.}

\usage{tree}

\format{이 자료는 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{volume}{부피(입방미터)}
    \item{diameter}{지름(cm). 지ㄹ듬은 지상에서 1.5m 지점에서 측정되었다.}
    \item{height}{높이(m)}
    }
}


\examples{
}

\references{
}
\source{
}
