% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diag.R
\name{vif}
\alias{print.summary.vif}
\alias{print.vif}
\alias{summary.vif}
\alias{vif}
\alias{vif.lm}
\title{Variance Inflation Factor and Collinearity Diagnostics}
\usage{
vif(mod, ...)

\method{vif}{lm}(mod, ...)

\method{print}{vif}(obj, ...)

\method{summary}{vif}(obj, ...)

\method{print}{summary.vif}(obj, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{mod}{an R object}

\item{...}{further arguments pass to or from other methods}
}
\description{
Variance Inflation Factor and Collinearity Diagnostics
}

