\encoding{utf8}
\name{waist}
\docType{data}
\alias{waist}

\title{waist}

\description{
과거에는 부모가 아이들의 옷을 구입할 때 바지 또는 치마의 허리부분을
아이의 목에 대어보고 치수를 어림짐작하곤 하였다. 아마 목둘레와
허리둘레가 비례한다는 것을 경험적으로 알고 전해진 행동이라고 생각된다.
자료 \code{waist}는 목둘레와 허리둘레가 어떠한 관계가 있는지를 알아보기
위해 대학생 16명을 무작위로 추출하여 조사한 자료이다.
}

\usage{waist}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{waist}{허리둘레(inch)}
    \item{neck}{목둘레(inch)}
    }
}

\references{
}

\source{
}

\keyword{datasets}