\encoding{utf8}
\name{wool}
\docType{data}
\alias{wool}

\title{
  양모의 세기
}

\description{
자료 \code{wool}는 Box & Cox의 1964년 논문에서 사용한 예제로, 양모의
강력을 알아보기 위해 \deqn(3^3) 요인실험을 수행한 결과이다.
}

\usage{wool}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{cycle}{반응변수. 시편이 끊어질 때까지의 측정 횟수.}
    \item{length}{시편의 길이}
    \item{load}{시편에 가한 하중}
    \item{amplitude}{하중을 가한 폭}
    }
}

\references{
}

\source{
}

\keyword{datasets}
