\encoding{utf8}
\name{wrapper}
\docType{data}
\alias{wrapper}

\title{
  wrapper
}

\description{
자료 \code{wrapper}는 빵 포장지의 강도를 높이기 위한 최적 조건을 찾기
위한 실험에서 얻어진 것이다. 

}

\usage{wrapper}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{y}{반응 변수. 포장지의 강도.}
    \item{x1}{밀봉 온도}
    \item{x2}{냉각 온도}
    \item{x3}{첨가되는 폴리에틸렌의 비율}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
