\encoding{utf8}
\name{yield}
\docType{data}
\alias{yield}

\title{
  yield
}

\description{
 자료 \code{yield}는 어느 화확공정에서 공정시간과 공정온도에 따른 효율을
 조사한 것이다.
}

\usage{yield}

\format{이 데이터프레임은 다음과 같은 열로 구성되어 있다.
  \describe{
    \item{yield}{효율}
    \item{time}{공정 시간}
    \item{temp}{공정 온도}
    }
}

\examples{
}

\references{
}

\source{
}

\keyword{datasets}
